+++ example/ios/Runner/Runner-Bridging-Header.h
@@ -1,1 +1,2 @@
 #import "GeneratedPluginRegistrant.h"
+#import <xinstall_flutter_plugin/XinstallFlutterPlugin.h>